requirements.txt
fastapi>=0.115.0
uvicorn[standard]
redis
aiohttp
websockets
pyyaml
python-multipart
pydantic>=2.0
python-dotenv
anthropic>=0.40.0
